Engine: Split Player.cs into Player.cs, ActionSet.cs
